+++ src/Services/organizationService.js
@@ -1,4 +1,4 @@
-import { get } from '../Services/privateApiService';
+import { get } from './privateApiService';
 
 const BASE_URL = "https://ongapi.alkemy.org/public/api";
 const BASE_ENDPOINT_ORGANIZATION = `/organization`;
@@ -7,5 +7,6 @@ const Id = 1;
 
 async function getDataOrganization() {
        return await get(`${BASE_URL}${BASE_ENDPOINT_ORGANIZATION}`, Id);
-    };
-  export { getDataOrganization };
+};
+
+export { getDataOrganization };
